requirements.txt
aiomysql==0.0.20
amqp==2.5.2
asgiref==3.2.7
atomicwrites==1.3.0
attrs==19.3.0
Automat==20.2.0
beautifulsoup4==4.9.0
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
cryptography==2.9
cssselect==1.1.0
cycler==0.10.0
dateparser==0.7.4
Django==3.0.5
fake-useragent==0.1.11
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.0
incremental==17.5.0
jieba==0.42.1
joblib==0.14.1
kiwisolver==1.2.0
kombu==4.6.8
lxml==4.5.0
matplotlib==3.2.1
more-itertools==8.2.0
mysqlclient==1.4.6
numpy==1.18.2
packaging==20.3
parsel==1.5.2
pluggy==0.13.1
ply==3.11
Protego==0.1.16
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pymongo==3.10.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytest==5.4.1
python-dateutil==2.8.1
pytz==2019.3
queuelib==1.5.0
redis==3.4.1
regex==2020.4.4
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
Scrapy==2.0.1
selenium==3.141.0
service-identity==18.1.0
six==1.14.0
sklearn==0.0
soupsieve==2.0
SQLAlchemy==1.3.16
sqlparse==0.3.1
Twisted==20.3.0
tzlocal==2.0.0
urllib3==1.25.8
vine==1.3.0
w3lib==1.21.0
wcwidth==0.1.9
xlrd==1.2.0
zipp==3.1.0
zope.interface==5.1.0
pandas==0.25.3